feat: deleteHighLight 기능 추가
- deleteHighLight 기능 개선

+++ BE/media/src/main/java/com/backtoback/media/video/service/VideoServiceImpl.java
@@ -254,9 +254,14 @@ public class VideoServiceImpl implements VideoService {
   @Override
   public void deleteHighLight(Long gameSeq){
     log.info("delete record");
-    Record record = recordRepository.findById(gameSeq.toString()).orElseThrow();
-    deleteFile(record.getRecordPath());
-    recordRepository.deleteById(gameSeq.toString());
+
+    Optional<Record> optionalRecord = recordRepository.findById(gameSeq.toString());
+
+    if(optionalRecord.isPresent()){
+      Record record = optionalRecord.get();
+      deleteFile(record.getRecordPath());
+      recordRepository.deleteById(gameSeq.toString());
+    }
 
     highLightRepository.findAll().forEach(highLight -> {
       log.info("delete highlight");
